modularize AddButton component
